+++ src/Oro/Bundle/UserBundle/Resources/config/services.yml
@@ -1,5 +1,5 @@
 parameters:
-    oro_user.manager.class:           Oro\Bundle\UserBundle\Entity\UserManager
+    oro_user.manager.class:           Pim\Bundle\UserBundle\Manager\UserManager
     oro_user.entity.class:            Pim\Bundle\UserBundle\Entity\User
     oro_user.group.entity.class:      Oro\Bundle\UserBundle\Entity\Group
     oro_user.role.entity.class:       Oro\Bundle\UserBundle\Entity\Role
